refactor(uart): 修改UART发送函数长度参数类型为uint32_t
将erdp_if_uart_send_bytes函数的长度参数从uint8_t改为uint32_t
以支持更大的数据包传输，同时更新循环计数器类型保持一致性

+++ Source/Interface/Hardware/inc/erdp_if_uart.h
@@ -78,7 +78,7 @@ void erdp_if_uart_init(ERDP_Uart_t uart, uint32_t baudrate, ERDP_UartMode_t mode
  * @param[in] len: Length of data to send (in bytes)
  * @return None
  */
-void erdp_if_uart_send_bytes(ERDP_Uart_t uart, const uint8_t* data, uint8_t len);
+void erdp_if_uart_send_bytes(ERDP_Uart_t uart, const uint8_t* data, uint32_t len);
 
 /**
  * @brief Read a single byte from specified UART
